fix(Register): Renamed component. Made it more clear what the component does when
